requirements.txt
python-telegram-bot==21.1
yt-dlp>=2025.01.01
python-dotenv==1.0.1
pycryptodome==3.20.0
